% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{AMI}
\alias{AMI}
\title{Adjusted Mutual Information}
\usage{
AMI(subject, query)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}
}
\description{
Adjusted Mutual Information
}
