% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{canonical_immune_markers}
\alias{canonical_immune_markers}
\title{Canonical marker genes of the hematopoietic lineage.}
\format{A data frame with 135 rows and 5 variables:
\describe{
  \item{gene_symbol}{Official gene symbol}
  \item{entrez_id}{Entrez Gene ID}
  \item{score}{Cell type score for this gene}
  \item{cell_code}{Coded cell type defined by marker genes}
  \item{cell_type}{Named cell type defined by marker genes}
}}
\source{
\url{https://www.cell.com/cell/abstract/S0092-8674(11)00005-5}
}
\usage{
canonical_immune_markers
}
\description{
A list of canonical immune cell marker genes compiled from Figure 1 in Novershtern et al., Cell (2011).
}
\keyword{datasets}
