% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{characterizeMarkerThresholds}
\alias{characterizeMarkerThresholds}
\title{Characterize Marker Thresholds}
\usage{
characterizeMarkerThresholds(FGDEGtab, esThresh = seq(0.5, 1, 0.01),
  nThresh = NULL, clusterColumn = "cluster_id", geneColumn = "entrez_id",
  effectColumn = "effect.size")
}
\arguments{
\item{FGDEGtab}{The raw FASTGenomics Differentially Expressed Gene table, e.g. the output from FASTGenomics calc_de_genes_nonparametric.}

\item{esThresh}{A vector of minimal effect sizes; defaults to \code{seq(0.5, 1, .01)}.}

\item{nThresh}{An optional vector of maximum cluster numbers that contain a potential marker gene; defaults to NULL, which takes the range between 1 and the maximum number of clusters found in the FGDGEtab.}

\item{clusterColumn}{Column name for cluster assignment; defaults to "cluster_id".}

\item{geneColumn}{Column name for gene ID; defaults to "entrez_id".}

\item{effectColumn}{Column name for effect size; defaults to "effect.size".}
}
\value{
The marker threshold matrix.
}
\description{
Construct a matrix of clusters with marker genes captured with
ranges of effect size and cluster hit threshold values.
The resulting matrix should help at selecting optimal threshold values.
}
