% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{clusterExclusivity}
\alias{clusterExclusivity}
\title{Cluster Exclusivity}
\usage{
clusterExclusivity(subject, query, maxIgnore = 0)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}

\item{maxIgnore}{Maximum count that is ignored in contingency matrix; defaults to 0.}
}
\value{
Vector of cluster exclusivities for each subject cluster.
}
\description{
Computes if query clusters are exclusively assigned to subject cluster.
}
