% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{clusterExprsSummaryMatrix}
\alias{clusterExprsSummaryMatrix}
\title{Construct a matrix of expression summary values per cluster.}
\usage{
clusterExprsSummaryMatrix(FGDEGtab, FGexprs, FGassign, Q = 0.5,
  cellColumn = "cell_id", geneColumn = "entrez_id",
  clusterColumn = "cluster_id")
}
\arguments{
\item{FGDEGtab}{The raw FASTGenomics Differentially Expressed Gene table, e.g. the output from FASTGenomics calc_de_genes_nonparametric.}

\item{FGexprs}{A sparse Matrix of expression values in cells x genes format.}

\item{FGassign}{data.frame with cluster assignments for cells.}

\item{Q}{Quantile used for expression value summary; defaults to 0.5, i.e. the median.}

\item{cellColumn}{Column name for cell name; defaults to "cell_id".}

\item{geneColumn}{Column name for gene ID; defaults to "entrez_id".}

\item{clusterColumn}{Column name for cluster assignment; defaults to "cluster_id".}
}
\value{
A gene x cluster matrix of summarized cluster expression values.
}
\description{
Construct a matrix of expression summary values per cluster.
}
