% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{clusterGranularity}
\alias{clusterGranularity}
\title{Cluster Granularity}
\usage{
clusterGranularity(subject, query, plot = FALSE)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}

\item{plot}{Display the query x subject matrix; defaults to \code{FALSE}.}
}
\value{
Vector of cluster granularities for each subject cluster.
}
\description{
Computes the proportion of query clusters assigned to each subject cluster.
}
