% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{clusterIntegrity}
\alias{clusterIntegrity}
\title{Cluster Integrity}
\usage{
clusterIntegrity(subject, query, maxIgnore = 0)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}

\item{maxIgnore}{Maximum count that is ignored in contingency matrix; defaults to 0.}
}
\value{
Vector of cluster integrity for each subject cluster.
}
\description{
Computes if query clusters are exclusively assigned to subject cluster.
}
