% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{confusionHeatmap}
\alias{confusionHeatmap}
\title{Plot a confusion heatmap.}
\usage{
confusionHeatmap(subject, query, logCounts = FALSE, rankCounts = FALSE, ...)
}
\arguments{
\item{subject}{\code{Vector} of reference cluster assignments.}

\item{query}{\code{Vector} of cluster assignments for comparison with reference.}

\item{logCounts}{Should counts in confusion matrix be \code{log10(counts+1)} transformed? Defaults to \code{FALSE}.}

\item{rankCounts}{Should counts in confusion matrix be ranked? Defaults to \code{FALSE}.}

\item{...}{Further parameters for \code{pheatmap} function.}
}
\value{
A \code{matrix} with possibly transformed counts.
}
\description{
Plot a confusion heatmap.
}
\examples{
a = c(rep("A", 1000), rep("B", 100), rep("C", 10))
b = c(rep("A", 500), rep("B", 595), rep("C", 15))
confusionHeatmap(a, b)
confusionHeatmap(a, b, logCounts = T)
confusionHeatmap(a, b, rankCounts = T)
}
