% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{EMI}
\alias{EMI}
\title{Expected Mutual Information}
\usage{
EMI(subject, query)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}
}
\value{
Expected Mutual Information, a numeric vector of length 1.
}
\description{
Expected Mutual Information
}
