% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{entropy}
\alias{entropy}
\title{Calculate Shannon Entropy}
\usage{
entropy(subject)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}
}
\value{
Mutual Information, a numeric vector of length 1.
}
\description{
Calculate Shannon Entropy
}
