% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{findMarkerThresholds}
\alias{findMarkerThresholds}
\title{Find Marker Thresholds}
\usage{
findMarkerThresholds(MTM, targetValue = max(MTM))
}
\arguments{
\item{MTM}{The marker threshold matrix.}

\item{targetValue}{The minimum target value that will be searched the the marker threshold matrix; defaults to \code{max(MTM)}.}
}
\value{
A list with elements minES and maxClustersPerGene.
}
\description{
Find optimized values for minES and maxClustersPerGene in marker threshold matrix.
Assumes that most stringent parameter combination is in the bottom left corner
and the least stringent combination is in the upper right corner of the
marker threshold matrix.
}
