% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{immune_markers_blueprint}
\alias{immune_markers_blueprint}
\title{Cell type marker genes and scores.}
\format{A data frame with 7,876 rows and 4 variables:
\describe{
  \item{gene_symbol}{Official gene symbol}
  \item{entrez_id}{Entrez Gene ID}
  \item{score}{Cell type score for this gene}
  \item{cell_type}{Named cell type defined by marker genes}
}}
\source{
\url{https://github.com/dviraran/SingleR/tree/master/data}
}
\usage{
immune_markers_blueprint
}
\description{
Marker gene scores generated from Blueprint+ENCODE expression profiles compiled for singleR package.
}
\keyword{datasets}
