% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{immune_markers_lm22}
\alias{immune_markers_lm22}
\title{Immune cell marker genes and scores.}
\format{A data frame with 12,034 rows and 4 variables:
\describe{
  \item{gene_symbol}{Official gene symbol}
  \item{entrez_id}{Entrez Gene ID}
  \item{score}{Cell type score for this gene}
  \item{cell_type}{Named cell type defined by marker genes}
}}
\source{
\url{https://doi.org/10.1038/nmeth.3337}
}
\usage{
immune_markers_lm22
}
\description{
The LM22 signature expression matrix used by Newman et al. (2015), with linear scale expression values divided by 1000 used as scores.
}
\keyword{datasets}
