% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{makeDonuts}
\alias{makeDonuts}
\title{Make Donut Plot.}
\usage{
makeDonuts(subject, query, subquery = NULL, colorList = NULL,
  pieLim = c(0, 4), pieCut = 2.5,
  piesPerRow = ceiling(sqrt(length(unique(subject)))), savePDF = FALSE)
}
\arguments{
\item{subject}{\code{Vector} of reference cluster assignments.}

\item{query}{\code{Vector} of cluster assignments for comparison with reference.}

\item{subquery}{\code{Vector} of lower level cluster assignments for comparison with reference. Defaults to \code{NULL}.}

\item{pieLim}{\code{vector} of length 2 giving the \code{xlim} values for pies; defaults to \code{c(0, 4)}.}

\item{pieCut}{\code{integer} giving the radius of the query pie if subquery is defined; should be within the \code{pieLim} range; defaults to \code{2.5}.}

\item{piesPerRow}{Number of pies per row.}

\item{colorList=NULL}{A named \code{list} of colors for clusters. The \code{list} needs elements "query" and optionally "subquery".}
}
\value{
A \code{list} with elements \code{donuts} (a ggplot object) and \code{data} (the underlying data.frame).
}
\description{
Make Donut Plot.
}
\examples{
add(1, 1)
add(10, 1)
}
