% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTGenomicsIO.R
\name{maxChunkFGH5}
\alias{maxChunkFGH5}
\title{Determine the number of chunks of a FASTGenomics Matrix HDF5 file.}
\usage{
maxChunkFGH5(filename, chunkSize = 10000)
}
\arguments{
\item{filename}{Name of FASTGenomics HDF5 file.}

\item{chunkSize}{Number of cells per chunk, defaults to 10000.}
}
\value{
The number of chunks.
}
\description{
Determine the number of chunks of a FASTGenomics Matrix HDF5 file.
}
