% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{newman2015}
\alias{newman2015}
\title{Immune cell marker genes and scores.}
\format{A data frame with 12034 rows and 4 variables:
\describe{
  \item{gene_symbol}{Official gene symbol}
  \item{entrez_id}{Entrez Gene ID}
  \item{score}{Cell type score for this gene}
  \item{cell_type}{Named cell type defined by marker genes}
}}
\source{
\url{https://doi.org/10.1038/nmeth.3337}
}
\usage{
newman2015
}
\description{
The log10 transformed expression matrix used in CIBERSORT to deconvolute immune cell types.
}
\keyword{datasets}
