% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{NMI}
\alias{NMI}
\title{Normalized Mutual Information}
\usage{
NMI(subject, query)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}
}
\description{
Normalized Mutual Information
}
