% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{novershtern2011}
\alias{novershtern2011}
\title{Marker genes of various immune cell types.}
\format{A data frame with 15088 rows and 3 variables:
\describe{
  \item{gene_id}{Entrez Gene ID}
  \item{score}{Cell type score for this gene}
  \item{cell_type}{Cell type defined by marker genes}
}}
\source{
\url{https://www.cell.com/cell/abstract/S0092-8674(11)00005-5}
}
\usage{
novershtern2011
}
\description{
A list of immune cell marker genes compiled from Novershtern et al., Cell (2011).
}
\keyword{datasets}
