% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTGenomicsIO.R
\name{observationsFGH5}
\alias{observationsFGH5}
\title{Retrieve observations from FASTGenomics Matrix HDF5 file.}
\usage{
observationsFGH5(filename)
}
\arguments{
\item{filename}{Name of FASTGenomics HDF5 file.}
}
\value{
A vector of observation IDs.
}
\description{
Observations typically correspond to cells.
}
