% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{plotDonuts}
\alias{plotDonuts}
\title{Plot donuts.}
\usage{
plotDonuts(subject, query, subquery = NULL, savePDF = FALSE, ...)
}
\arguments{
\item{subject}{\code{Vector} of reference cluster assignments.}

\item{query}{\code{Vector} of cluster assignments for comparison with reference.}

\item{subquery}{\code{Vector} of lower level cluster assignments for comparison with reference. Defaults to \code{NULL}.}

\item{savePDF}{Should plots be saved in PDF? Defaults to \code{FALSE}.}

\item{...}{Further parameters used for function \code{makeDonuts}.}
}
\description{
Plot donuts.
}
