% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{plotLegend}
\alias{plotLegend}
\title{Plot the legend(s) for donut plots.}
\usage{
plotLegend(data, subquery = FALSE)
}
\arguments{
\item{data}{\code{data.frame} returned from \code{makeDonuts} in list element \code{data}.}
}
\description{
Plot the legend(s) for donut plots.
}
