% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTGenomicsIO.R
\name{readAEoutput}
\alias{readAEoutput}
\title{Read FASTGenomics Autoencoder Output}
\usage{
readAEoutput(filename = "encoded_data.h5")
}
\arguments{
\item{filename}{Name of Autoencoder output file, defaults to \code{encoded_data.h5}.}
}
\description{
Read FASTGenomics Autoencoder Output
}
