% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClEvaR.R
\name{readFGH5}
\alias{readFGH5}
\title{Read expression data from FASTGenomics HDF5 file.}
\usage{
readFGH5(fileName, cellIDs = NULL, geneIDs = NULL, matrixName = "/matrix",
  observationsName = "/obs_names", variablesName = "/var_names")
}
\arguments{
\item{fileName}{Name of FASTGenomics HDF5 file.}

\item{cellIDs}{Optional vector of cell IDs (i.e. row names) to be read from FASTGenomics HDF5 file.}

\item{geneIDs}{Optional vector of gene IDs (i.e. column names) to be read from FASTGenomics HDF5 file.}

\item{matrixName}{Dataset name of expression matrix in HDF5 file; defaults to "/matrix".}

\item{observationsName}{Dataset name of observation (i.e. cell) name vector in HDF5 file; defaults to "/obs_names".}

\item{variablesName}{Dataset name of variable (i.e. gene) name vector in HDF5 file; defaults to "/var_names".}
}
\value{
A sparse Matrix of expression values in cells x genes format.
}
\description{
Read expression data from FASTGenomics HDF5 file.
}
