% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{scale_01}
\alias{scale_01}
\title{Scale to 0 - 1 range}
\usage{
scale_01(x, invert = TRUE)
}
\arguments{
\item{x}{Numeric vector of values to be scaled}

\item{invert}{Scales smallest value in x to 1 and largest to 0 if \code{TRUE} and to 0 and 1 respectively if \code{FALSE}. Defaults to \code{TRUE}.}
}
\value{
Numeric vector scaled of values.
}
\description{
Scale to 0 - 1 range
}
