% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{scoreCellsInChunks}
\alias{scoreCellsInChunks}
\title{Classify cells using marker genes.}
\usage{
scoreCellsInChunks(dataset = NULL, filename = NULL,
  chunkSize = 10000, markerGenes = NULL, geneColumn = "entrez_id",
  scoreColumn = "score", classColumn = "cell_type", verbose = TRUE)
}
\arguments{
\item{dataset}{A FASTGenomics Matrix of expression values.}

\item{filename}{Name of FASTGenomics HDF5 file.}

\item{markerGenes}{An optional data frame with scores of marker genes for cell types; the immune cell markers compiled from Newman et al., Nature Methods (2015) is provided as default.}

\item{geneColumn}{Gene column name in markerGenes data frame; defaults to \code{entrez_id}.}

\item{scoreColumn}{Score column name in markerGenes data frame; defaults to \code{score}.}

\item{classColumn}{Class column name in markerGenes data frame; defaults to \code{cell_type}.}
}
\value{
A data frame with class assignment and scores for classes per cell.
}
\description{
Performs cell typing on chunks of FASTGenomics expression files. Useful for large datasets and limited computing ressources.
}
