% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTGenomicsIO.R
\name{variablesFGH5}
\alias{variablesFGH5}
\title{Retrieve variables from FASTGenomics Matrix HDF5 file.}
\usage{
variablesFGH5(filename)
}
\arguments{
\item{filename}{Name of FASTGenomics HDF5 file.}
}
\value{
A vector of variable IDs.
}
\description{
Variables typically correspond to genes.
}
