% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{zeisel2018}
\alias{zeisel2018}
\title{Molecular architecture of the mouse nervous system.}
\format{A list with 3 items:
\describe{
  \item{cell_types}{A data frame describing the 265 nervous cell types.}
  \item{cell_markes}{A data frame with marker genes for cell types as symbols and entrez ids.}
  \item{cell_metadata}{A data frame with metadata of 160k cells from the mouse nervous tissue.}
}}
\source{
\url{http://mousebrain.org}
}
\usage{
zeisel2018
}
\description{
Molecular architecture of the mouse nervous system.
}
\keyword{datasets}
