% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{integrityPlot}
\alias{integrityPlot}
\title{Generate cluster integrity plot.}
\usage{
integrityPlot(subject, query)
}
\arguments{
\item{subject}{\code{Vector} of reference cluster assignments.}

\item{query}{\code{Vector} of cluster assignments for comparison with reference.}
}
\value{
A \code{ggplot} object.
}
\description{
A Granularity vs. Integrity scatterplot.
}
\examples{
a = c(rep("A", 1000), rep("B", 100), rep("C", 10))
b = c(rep("A", 500), rep("B", 595), rep("C", 15))
integrityPlot(a, b)

data(zeisel2018)
integrityPlot(subject = zeisel2018$cell_metadata$class,
              query = zeisel2018$cell_metadata$cluster_name)
}
