% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{MI}
\alias{MI}
\title{Mutual Information}
\usage{
MI(subject, query)
}
\arguments{
\item{subject}{Vector of reference cluster assignments.}

\item{query}{Vector of cluster assignments for comparison.}
}
\value{
Mutual Information, a numeric vector of length 1.
}
\description{
Mutual Information
}
