% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTGenomicsIO.R
\name{readFGH5}
\alias{readFGH5}
\title{Read FASTGenomics Matrix in HDF5 format.}
\usage{
readFGH5(filename, chunk = NULL, chunkSize = 10000, fromCell = NULL,
  toCell = NULL)
}
\arguments{
\item{filename}{Name of FASTGenomics HDF5 file.}

\item{chunk}{Optional 1-based number of chunk to read.}

\item{chunkSize}{Number of cells per chunk, defaults to 10000.}

\item{fromCell}{Optional 1-based index of first cell to read.}

\item{toCell}{Optional 1-based index of last cell to read.}
}
\value{
A sparse \code{Matrix} of expression values.
}
\description{
Read FASTGenomics Matrix in HDF5 format.
}
