% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{scoreCells}
\alias{scoreCells}
\title{Classify cells using marker genes.}
\usage{
scoreCells(dataset, markerGenes = NULL, geneColumn = "entrez_id",
  scoreColumn = "score", classColumn = "cell_type")
}
\arguments{
\item{dataset}{A FASTGenomics Matrix of expression values.}

\item{markerGenes}{An optional data frame with scores of marker genes for cell types; the canonical immune cell markers compiled from Novashtern et al., Cell (2011) is provided as default.}

\item{geneColumn}{Gene column name in markerGenes data frame; defaults to \code{entrez_id}.}

\item{scoreColumn}{Score column name in markerGenes data frame; defaults to \code{score}.}

\item{classColumn}{Class column name in markerGenes data frame; defaults to \code{cell_type}.}
}
\value{
A data frame with class assignment and scores for classes per cell.
}
\description{
Classify cells using marker genes.
}
